refactor: #123 - response 구조 분리 및 validation 의존성 정리
