+++ src/features/auth/signup/lib/mapSignupValidationErrors.ts
@@ -1,9 +1,8 @@
 import type { ZodError } from "zod";
 
-import type { ValidationError } from "@/features/auth/lib/response";
 import { VALIDATION_REASON } from "@/lib/validation/reasons";
-
-import { ValidationReason } from "../../../../lib/validation/validation.types";
+import { ValidationReason } from "@/lib/validation/validation.types";
+import { ValidationError } from "@/lib/validation/validation-error.types";
 
 /**
  * Zod issue.path를 따라 원본 input에서 실제 값을 조회하는 유틸 함수
